WIP - ca compile- re arranger - on passe pas la grid mais direct l'array de Point a Champ - conflict resolu
