+++ src/include/point_bis.hpp
@@ -1,0 +1,17 @@
+#ifndef _point_hpp
+#define _point_hpp
+
+class Point{
+
+	private :
+		double x_ , y_ ;
+	public:
+		Point(){x_ = 0. ; y_ = 0. ;}
+		~Point(){};
+		double getX(){return x_;};
+		double getY(){return y_;};
+		void setX(double x){x_ = x ;};
+		void setY(double y){y_ = y ;};
+};
+
+#endif // _point_hpp
